Users can now use brackets to specify the title for their gist.
Pick up the content inside the first [] as the snippet thumbnail's
title. For example, "[Apple is delicious] It's affordable and healthy."
will pick up "Apple is delicious" as the title. If no brackets are
found,  it shows to the original description. It provides users the
flexibility to decide what to be shown in the thumbnail.