bugs fixed
1. The raw modal’s style is not in effect
2. Occasional errors complaining about the null mainWindow when sending
the key events. This occurs when the mainWindow is destroyed but the
electronLocalshortcut is still listening the key event and tries to
call mainWindow.webContents to emit events. (In macOS, the app won’t
quit when the window is close, which means the listener is not
unregistered, maybe we should call it leaking)