Update to webpack 2.0
What’s new in webpack 2?

https://news.ycombinator.com/item?id=11017595
“Huge. Between webpack and Babel, JS is quickly progressing towards
becoming a powerful language that is easy to write, read and reason
about. I'm currently in the process of moving an inherited "minified by
hand" codebase to ES6, and the difference is incredible. Not only are
linters catching hundreds of previously unknown edge cases and bugs,
but the end result is more performant and smaller in file size. It's
very satisfying, makes me not mind coding in JS again.”