Optimize the speed for submitting the edit form
In the past, we close the dialog after getting response from the server
both for the “create” and “edit” actions. However, the state change for
the “edit” action is less intensive than “create”, for example, the
active language tag and the active gist are likely be the same as
before. Therefore, we decide the close the dialog without waiting for
the server’s response, which provides better user experience.